refactor database queries and express routes
